move observation methods to separate module
